# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BEYOND-MOBILE-BACKEND",
        "title": "BEYOND-MOBILE-BACKEND",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/internal/v1/articles/delete": {
            "post": {
                "description": "Remove New Articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Remove Articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Remove Articles Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.DeleteArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RecordNotFoundArticles"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/internal/v1/articles/insert": {
            "post": {
                "description": "Add New Articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "New Articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Insert Articles Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.NewArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/internal/v1/articles/list": {
            "post": {
                "description": "Inquiry List Articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Lit Articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Select List Articles Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.SelectListArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.SuccessListArticles"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RecordNotFoundArticles"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/internal/v1/articles/update": {
            "patch": {
                "description": "Update Existing Articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Update Articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update Articles Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.UpdateArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.BasicSuccess"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RecordNotFoundArticles"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
        "/api/v1/articles/list": {
            "post": {
                "description": "Inquiry List Articles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Articles"
                ],
                "summary": "Lit Articles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "X-NONCE",
                        "name": "X-NONCE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-SIGNATURE",
                        "name": "X-SIGNATURE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-DEVICE-ID",
                        "name": "X-DEVICE-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-TIMESTAMP",
                        "name": "X-TIMESTAMP",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LATITUDE",
                        "name": "X-LATITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "X-LONGITUDE",
                        "name": "X-LONGITUDE",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "List Articles Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_request.SelectListArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.SuccessListArticles"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.DeferenceDeviceArtice"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.RecordNotFoundArticles"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/backend-mobile-api_model_dto_swagger.CommonError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "backend-mobile-api_model_dto_request.Article": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "active_after_day": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "backend-mobile-api_model_dto_request.ArticleByList": {
            "type": "object",
            "properties": {
                "active_after_day": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.DeleteArticleRequest": {
            "type": "object",
            "required": [
                "article",
                "device_id"
            ],
            "properties": {
                "article": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.Article"
                },
                "device_id": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.NewArticle": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "active_after_day": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "backend-mobile-api_model_dto_request.NewArticleRequest": {
            "type": "object",
            "required": [
                "device_id",
                "new_article"
            ],
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "new_article": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.NewArticle"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.SelectListArticleRequest": {
            "type": "object",
            "required": [
                "article",
                "device_id"
            ],
            "properties": {
                "article": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.ArticleByList"
                },
                "device_id": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_request.UpdateArticleRequest": {
            "type": "object",
            "required": [
                "article",
                "device_id"
            ],
            "properties": {
                "article": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_request.Article"
                },
                "device_id": {
                    "type": "string"
                }
            }
        },
...
        "backend-mobile-api_model_dto_response.ArticleResponse": {
            "type": "object",
            "properties": {
                "active_after_day": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "backend-mobile-api_model_dto_response.SelectArticleResponse": {
            "type": "object",
            "properties": {
                "articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/backend-mobile-api_model_dto_response.ArticleResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.BasicSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "string",
                    "example": "00"
                }
            }
        },
        "backend-mobile-api_model_dto_swagger.CommonError": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": "time out"
                },
                "message": {
                    "type": "string",
                    "example": "server is busy"
                },
                "status_code": {
                    "type": "string",
                    "example": "119"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.DeferenceDeviceArtice": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "deference device"
                },
                "status_code": {
                    "type": "string",
                    "example": "171"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.RecordNotFoundArticles": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "record not found"
                },
                "status_code": {
                    "type": "string",
                    "example": "160"
                }
            }
        },
...
        "backend-mobile-api_model_dto_swagger.SuccessListArticles": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/backend-mobile-api_model_dto_response.SelectArticleResponse"
                },
                "error": {
                    "type": "string",
                    "example": ""
                },
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "string",
                    "example": "00"
                }
            }
        },
...
    }
}